refactor(ui): consolidate getAppInfo() consumers onto shared cached hook
Generalize the cached desktop-version hook into useAppInfo(), which owns the
single module-level getAppInfo() fetch and shares the full AppInfo across every
consumer. useDesktopVersion is now a thin wrapper returning info?.version, and
its old module keeps re-exporting it so existing imports stay valid.

Migrate the four ad-hoc getAppInfo() consumers off their own
useEffect(() => getAppInfo().then(...)) blocks:
- Settings/AboutSettings
- Settings/SettingsPanel
- Settings/UpdateSettings
- StatusBar/UpdateIndicator
- UpdateNotification (uses the version-only useDesktopVersion wrapper)

getAppInfo() is now invoked at most once per session regardless of how many of
these mount. No user-facing behavior change (develop badge, version display,
update dot all identical).

Closes #1407
